chore: restructured project
